api/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
dotenv==0.9.9
pydantic==2.12.5
python-multipart==0.0.21
google-cloud-storage==3.7.0
python-magic==0.4.27
langgraph==1.0.5
langchain==1.2.0
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-qdrant==1.1.0
langchain-community==1.0.0a1
qdrant-client==1.16.2
